Add nix flake package (#5)
Both native AOT and ReadyToRun are broken because they rely on binaries
downloaded elsewhere, and Nix doesn't really like to run those in it's
sandboxed environment. Thus if you're a nix user you're only getting a
JIT binary, which might be slower but not by a lot.

In other news, try to keep nuget.json up to date when you're changing
any of the lockfiles to prevent breakages. And update the version number
because the dotnet builder doesn't like setting "unstable" as a version
number.

But other than those two problems this should *just work*.